requirements.txt
streamlit 
nltk 
groq 
python-dotenv